+++ lib/src/model/asn1/ASN1ObjectType.dart
@@ -1,0 +1,1 @@
+enum ASN1ObjectType { PRIM, CONST }